requirements.txt
flask
flask-restful
flask-cors
pyyaml
gevent
flask-api
requests
lxml